Overload set/getData in order to store C++ strings

+++ wrappers/cpp/object.hh
@@ -45,13 +45,15 @@ namespace linphone {
 	public:
 		template <class T> void setData(const std::string &key, const std::shared_ptr<T> &data) {
 			std::shared_ptr<T> *newSharedPtr = new std::shared_ptr<T>(data);
-			belle_sip_object_data_set((::belle_sip_object_t *)mPrivPtr, key.c_str(), newSharedPtr, deleteSharedPtr<T>);
+			belle_sip_object_data_set(mPrivPtr, key.c_str(), newSharedPtr, deleteSharedPtr<T>);
 		}
 		template <class T> std::shared_ptr<T> getData(const std::string &key) const {
 			void *dataPtr = belle_sip_object_data_get((::belle_sip_object_t *)mPrivPtr, key.c_str());
 			if (dataPtr == NULL) return nullptr;
 			else return *dynamic_cast<T*>(dataPtr);
 		}
+		void setData(const std::string &key, const std::string &data);
+		const std::string &getData(const std::string &key) const;
 	
 	protected:
 		template <class T>
@@ -97,9 +99,8 @@ namespace linphone {
 		static std::list<std::string> cStringArrayToCppList(const char **cArray);
 	
 	private:
-		template <class T> static void deleteSharedPtr(std::shared_ptr<T> *ptr) {
-			delete ptr;
-		}
+		template <class T> static void deleteSharedPtr(std::shared_ptr<T> *ptr) {if (ptr != NULL) delete ptr;}
+		static void deleteString(std::string *str) {if (str != NULL) delete str;}
 	
 	protected:
 		::belle_sip_object_t *mPrivPtr;
